test: 좋아요 관련 로직 테스트코드 이동
