+++ src/test/java/com/poortorich/chat/facade/ChatFacadeTest.java
@@ -5,20 +5,17 @@ import com.poortorich.chat.entity.Chatroom;
 import com.poortorich.chat.entity.enums.ChatroomRole;
 import com.poortorich.chat.entity.enums.RankingStatus;
 import com.poortorich.chat.request.ChatroomCreateRequest;
-import com.poortorich.chat.request.ChatroomLikeUpdateRequest;
 import com.poortorich.chat.request.enums.SortBy;
 import com.poortorich.chat.response.AllChatroomsResponse;
 import com.poortorich.chat.response.ChatroomCoverInfoResponse;
 import com.poortorich.chat.response.ChatroomCreateResponse;
 import com.poortorich.chat.response.ChatroomDetailsResponse;
 import com.poortorich.chat.response.ChatroomInfoResponse;
-import com.poortorich.chat.response.ChatroomLikeStatusResponse;
 import com.poortorich.chat.response.ChatroomRoleResponse;
 import com.poortorich.chat.response.ChatroomsResponse;
 import com.poortorich.chat.service.ChatMessageService;
 import com.poortorich.chat.service.ChatParticipantService;
 import com.poortorich.chat.service.ChatroomService;
-import com.poortorich.like.service.LikeService;
 import com.poortorich.s3.service.FileUploadService;
 import com.poortorich.tag.service.TagService;
 import com.poortorich.user.entity.User;
@@ -56,8 +53,6 @@ class ChatFacadeTest {
     private TagService tagService;
     @Mock
     private ChatMessageService chatMessageService;
-    @Mock
-    private LikeService likeService;
 
     @InjectMocks
     private ChatFacade chatFacade;
@@ -244,46 +239,6 @@ class ChatFacadeTest {
         assertThat(response.getHasPassword()).isTrue();
         assertThat(response.getHostProfile().getUserId()).isEqualTo(user.getId());
         assertThat(response.getHostProfile().getIsHost()).isTrue();
-    }
-
-    @Test
-    @DisplayName("채팅방 좋아요 상태 조회 성공")
-    void getChatroomLikeSuccess() {
-        String username = "testUser";
-        User user = User.builder().username(username).build();
-
-        when(userService.findUserByUsername(username)).thenReturn(user);
-        when(chatroomService.findById(chatroomId)).thenReturn(chatroom);
-        when(likeService.getLikeStatus(user, chatroom)).thenReturn(true);
-        when(likeService.getLikeCount(chatroom)).thenReturn(3L);
-
-        ChatroomLikeStatusResponse result = chatFacade.getChatroomLike(username, chatroomId);
-
-        assertThat(result).isNotNull();
-        assertThat(result.getIsLiked()).isTrue();
-        assertThat(result.getLikeCount()).isEqualTo(3L);
-    }
-
-    @Test
-    @DisplayName("좋아요 상태 변경 성공")
-    void updateChatroomLikeSuccess() {
-        String username = "testUser";
-        Long chatroomId = 1L;
-        User user = User.builder().username(username).build();
-        Chatroom chatroom = Chatroom.builder().id(chatroomId).build();
-        ChatroomLikeUpdateRequest request = new ChatroomLikeUpdateRequest(true);
-
-        when(userService.findUserByUsername(username)).thenReturn(user);
-        when(chatroomService.findById(chatroomId)).thenReturn(chatroom);
-        when(likeService.getLikeStatus(user, chatroom)).thenReturn(true);
-        when(likeService.getLikeCount(chatroom)).thenReturn(3L);
-
-        ChatroomLikeStatusResponse result = chatFacade.updateChatroomLike(username, chatroomId, request);
-
-        verify(likeService).updateLikeStatus(user, chatroom, request.getIsLiked());
-        assertThat(result).isNotNull();
-        assertThat(result.getIsLiked()).isTrue();
-        assertThat(result.getLikeCount()).isEqualTo(3L);
     }
 
     @Test
